src/lemniscat/plugin/filetransform/requirements.txt
lemniscat.core>=0.2.1
PyYAML==6.0.1
pytz==2023.3
python-hcl2>=4.3.0
